+++ src/FlabIt.Guardians/GenericGuardiansExtension.cs
@@ -19,24 +19,27 @@ namespace FlabIt.Guardians
         }
 
         /// <summary>
-        /// Throws an <see cref="ArgumentNullException"/> when the specified <paramref name="argument"/> is null.
+        /// Throws an <see cref="ArgumentNullException"/> when the specified <paramref name="argument"/> is null.<br />
+        /// When <paramref name="argument" /> is not null, it will be returned.
         /// </summary>
         /// <typeparam name="TArgument">The type of <paramref name="argument"/>.</typeparam>
         /// <param name="argument">The argument.</param>
         /// <param name="argumentName">The name of the argument that, when specified, will be used instead of the default one.</param>
         /// <param name="message">A custom message that, when specified, will be used instead of the default one.</param>
         /// <exception cref="ArgumentNullException">Raised when <paramref name="argument"/> is null.</exception>
+        /// <returns>The <paramref name="argument" /> when it is not null.</returns>
         [ContractAnnotation("argument:null => halt")]
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static void ThrowIfNull<TArgument>(
-            [CanBeNull, NoEnumeration, ValidatedNotNull] this TArgument argument,
-            [CanBeNull, InvokerParameterName] string argumentName = null,
-            [CanBeNull] string message = null)
+        [NotNull]
+        public static TArgument ThrowIfNull<TArgument>(
+            [System.Diagnostics.CodeAnalysis.NotNull, CanBeNull, NoEnumeration, ValidatedNotNull] this TArgument? argument,
+            [CanBeNull, InvokerParameterName] string? argumentName = null,
+            [CanBeNull] string? message = null)
             where TArgument : class
         {
             if (!(argument is null))
-                return;
+                return argument;
 
             argumentName ??= nameof(argument);
 
@@ -56,10 +59,11 @@ namespace FlabIt.Guardians
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         [NotNull]
+        [Obsolete("Will be replaced by 'ThrowIfNull()' in upcoming versions.")]
         public static TArgument PassThroughNonNull<TArgument>(
-            [CanBeNull, NoEnumeration, ValidatedNotNull] this TArgument argument,
-            [CanBeNull, InvokerParameterName] string argumentName = null,
-            [CanBeNull] string message = null)
+            [System.Diagnostics.CodeAnalysis.NotNull, CanBeNull, NoEnumeration, ValidatedNotNull] this TArgument? argument,
+            [CanBeNull, InvokerParameterName] string? argumentName = null,
+            [CanBeNull] string? message = null)
             where TArgument : class
         {
             if (!(argument is null))