+++ src/FlabIt.Guardians/StringGuardiansExtension.cs
@@ -15,24 +15,24 @@ namespace FlabIt.Guardians
     public static class StringGuardiansExtension
     {
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static bool IsEmpty([NotNull] string argument)
+        private static bool IsEmpty(string argument)
         {
             return argument.Length == 0;
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static bool IsShorterThan([NotNull] string argument, int length)
+        private static bool IsShorterThan(string argument, int length)
         {
             return argument.Length < length;
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static bool IsLargerThan([NotNull] string argument, int length)
+        private static bool IsLargerThan(string argument, int length)
         {
             return argument.Length > length;
         }
 
-        private static bool IsOnlyWhitespaces([NotNull] string argument)
+        private static bool IsOnlyWhitespaces(string argument)
         {
             if (argument.Length == 0)
                 return false;
@@ -46,22 +46,22 @@ namespace FlabIt.Guardians
             return true;
         }
 
-        private static string GetIsEmptyErrorMessage([NotNull] string argumentName)
+        private static string GetIsEmptyErrorMessage(string argumentName)
         {
             return string.Format(CultureInfo.InvariantCulture, Resources.Exception_ArgumentOfTypeXEmptyMessageWithParamName, argumentName, typeof(string).FullName);
         }
 
-        private static string GetIsShorterThanErrorMessage([NotNull] string argumentName, int expectedLength, int actualLength)
+        private static string GetIsShorterThanErrorMessage(string argumentName, int expectedLength, int actualLength)
         {
             return string.Format(CultureInfo.InvariantCulture, Resources.Exception_ArgumentOfTypeXShorterThanMessageWithParamName, argumentName, typeof(string).FullName, expectedLength, actualLength);
         }
 
-        private static string GetIsLargerThanErrorMessage([NotNull] string argumentName, int expectedLength, int actualLength)
+        private static string GetIsLargerThanErrorMessage(string argumentName, int expectedLength, int actualLength)
         {
             return string.Format(CultureInfo.InvariantCulture, Resources.Exception_ArgumentOfTypeXLargerThanMessageWithParamName, argumentName, typeof(string).FullName, expectedLength, actualLength);
         }
 
-        private static string GetIsOnlyWhitespacesErrorMessage([NotNull] string argumentName)
+        private static string GetIsOnlyWhitespacesErrorMessage(string argumentName)
         {
             return string.Format(CultureInfo.InvariantCulture, Resources.Exception_ArgumentOfTypeXOnlyWhitespaceMessageWithParamName, argumentName, typeof(string).FullName);
         }
@@ -76,18 +76,20 @@ namespace FlabIt.Guardians
         /// <param name="message">A custom message that, when specified, will be used instead of the default one.</param>
         /// <exception cref="ArgumentNullException">Raised when <paramref name="argument"/> is null.</exception>
         /// <exception cref="ArgumentEmptyException">Raised when <paramref name="argument"/> is empty.</exception>
+        /// <returns>The <paramref name="argument" /> when it is not null or empty.</returns>
         [ContractAnnotation("argument:null => halt")]
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static void ThrowIfNullOrEmpty(
-            [CanBeNull, ValidatedNotNull] this string argument,
-            [CanBeNull, InvokerParameterName] string argumentName = null,
-            [CanBeNull] string message = null)
+        [NotNull]
+        public static string ThrowIfNullOrEmpty(
+            [System.Diagnostics.CodeAnalysis.NotNull, CanBeNull, ValidatedNotNull] this string? argument,
+            [CanBeNull, InvokerParameterName] string? argumentName = null,
+            [CanBeNull] string? message = null)
         {
-            argument.ThrowIfNull(argumentName, message);
+            argument = argument.ThrowIfNull(argumentName, message);
 
             if (!IsEmpty(argument))
-                return;
+                return argument;
 
             argumentName ??= nameof(argument);
 
@@ -107,12 +109,13 @@ namespace FlabIt.Guardians
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         [NotNull]
+        [Obsolete("Will be replaced by 'ThrowIfNullOrEmpty()' in upcoming versions.")]
         public static string PassThroughNonNullNorEmpty(
-            [CanBeNull, ValidatedNotNull] this string argument,
-            [CanBeNull, InvokerParameterName] string argumentName = null,
-            [CanBeNull] string message = null)
+            [System.Diagnostics.CodeAnalysis.NotNull, CanBeNull, ValidatedNotNull] this string? argument,
+            [CanBeNull, InvokerParameterName] string? argumentName = null,
+            [CanBeNull] string? message = null)
         {
-            argument.ThrowIfNull(argumentName, message);
+            argument = argument.ThrowIfNull(argumentName, message);
 
             if (!IsEmpty(argument))
                 return argument;
@@ -133,16 +136,19 @@ namespace FlabIt.Guardians
         /// <exception cref="ArgumentNullException">Raised when <paramref name="argument"/> is null.</exception>
         /// <exception cref="ArgumentEmptyException">Raised when <paramref name="argument"/> is empty.</exception>
         /// <exception cref="ArgumentWhitespaceException">Raised when <paramref name="argument"/> consists only of whitespaces.</exception>
+        /// <returns>The <paramref name="argument" /> when it is not null nor empty nor consists only of whitespaces.</returns>
         [ContractAnnotation("argument:null => halt")]
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static void ThrowIfNullOrEmptyOrWhitespace(
-            [CanBeNull, ValidatedNotNull] this string argument,
-            [CanBeNull, InvokerParameterName] string argumentName = null,
-            [CanBeNull] string message = null)
+        [NotNull]
+        public static string ThrowIfNullOrEmptyOrWhitespace(
+            [System.Diagnostics.CodeAnalysis.NotNull, CanBeNull, ValidatedNotNull] this string? argument,
+            [CanBeNull, InvokerParameterName] string? argumentName = null,
+            [CanBeNull] string? message = null)
         {
-            argument.ThrowIfNullOrEmpty(argumentName, message);
-            argument.ThrowIfNullOrWhitespace(argumentName, message);
+            return argument
+                .ThrowIfNullOrEmpty(argumentName, message)
+                .ThrowIfNullOrWhitespace(argumentName, message);
         }
 
         /// <summary>
@@ -159,10 +165,11 @@ namespace FlabIt.Guardians
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         [NotNull]
+        [Obsolete("Will be replaced by 'ThrowIfNullOrEmptyOrWhitespace()' in upcoming versions.")]
         public static string PassThroughNonNullNorEmptyNorWhitespace(
-            [CanBeNull, ValidatedNotNull] this string argument,
-            [CanBeNull, InvokerParameterName] string argumentName = null,
-            [CanBeNull] string message = null)
+            [System.Diagnostics.CodeAnalysis.NotNull, CanBeNull, ValidatedNotNull] this string? argument,
+            [CanBeNull, InvokerParameterName] string? argumentName = null,
+            [CanBeNull] string? message = null)
         {
             argument.ThrowIfNullOrEmpty(argumentName, message);
 
@@ -178,18 +185,20 @@ namespace FlabIt.Guardians
         /// <param name="message">A custom message that, when specified, will be used instead of the default one.</param>
         /// <exception cref="ArgumentNullException">Raised when <paramref name="argument"/> is null.</exception>
         /// <exception cref="ArgumentWhitespaceException">Raised when <paramref name="argument"/> consists only of whitespaces.</exception>
+        /// <returns>The <paramref name="argument" /> when it is not null nor consists only of whitespaces.</returns>
         [ContractAnnotation("argument:null => halt")]
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static void ThrowIfNullOrWhitespace(
-            [CanBeNull, ValidatedNotNull] this string argument,
-            [CanBeNull, InvokerParameterName] string argumentName = null,
-            [CanBeNull] string message = null)
+        [NotNull]
+        public static string ThrowIfNullOrWhitespace(
+            [System.Diagnostics.CodeAnalysis.NotNull, CanBeNull, ValidatedNotNull] this string? argument,
+            [CanBeNull, InvokerParameterName] string? argumentName = null,
+            [CanBeNull] string? message = null)
         {
-            argument.ThrowIfNull(argumentName, message);
+            argument = argument.ThrowIfNull(argumentName, message);
 
             if (!IsOnlyWhitespaces(argument))
-                return;
+                return argument;
 
             argumentName ??= nameof(argument);
 
@@ -209,12 +218,13 @@ namespace FlabIt.Guardians
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         [NotNull]
+        [Obsolete("Will be replaced by 'ThrowIfNullOrWhitespace()' in upcoming versions.")]
         public static string PassThroughNonNullNorWhitespace(
-            [CanBeNull, ValidatedNotNull] this string argument,
-            [CanBeNull, InvokerParameterName] string argumentName = null,
-            [CanBeNull] string message = null)
+            [System.Diagnostics.CodeAnalysis.NotNull, CanBeNull, ValidatedNotNull] this string? argument,
+            [CanBeNull, InvokerParameterName] string? argumentName = null,
+            [CanBeNull] string? message = null)
         {
-            argument.ThrowIfNull(argumentName, message);
+            argument = argument.ThrowIfNull(argumentName, message);
 
             if (!IsOnlyWhitespaces(argument))
                 return argument;
@@ -234,19 +244,21 @@ namespace FlabIt.Guardians
         /// <param name="message">A custom message that, when specified, will be used instead of the default one.</param>
         /// <exception cref="ArgumentNullException">Raised when <paramref name="argument"/> is null.</exception>
         /// <exception cref="ArgumentLengthShorterThanException">Raised when <paramref name="argument"/> is shorter than <paramref name="length"/>.</exception>
+        /// <returns>The <paramref name="argument" /> when it has the same or greater length than specified with <paramref name="length"/>.</returns>
         [ContractAnnotation("argument:null => halt")]
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static void ThrowIfShorterThan(
-            [NotNull, ValidatedNotNull] this string argument,
+        [NotNull]
+        public static string ThrowIfShorterThan(
+            [System.Diagnostics.CodeAnalysis.NotNull, NotNull, ValidatedNotNull] this string argument,
             int length,
-            [CanBeNull, InvokerParameterName] string argumentName = null,
-            [CanBeNull] string message = null)
+            [CanBeNull, InvokerParameterName] string? argumentName = null,
+            [CanBeNull] string? message = null)
         {
-            argument.ThrowIfNull(argumentName, message);
+            argument = argument.ThrowIfNull(argumentName, message);
 
             if (!IsShorterThan(argument, length))
-                return;
+                return argument;
 
             argumentName ??= nameof(argument);
 
@@ -267,11 +279,12 @@ namespace FlabIt.Guardians
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         [NotNull]
+        [Obsolete("Will be replaced by 'ThrowIfShorterThan()' in upcoming versions.")]
         public static string PassThroughNonShorterThan(
-            [NotNull, ValidatedNotNull] this string argument,
+            [System.Diagnostics.CodeAnalysis.NotNull, NotNull, ValidatedNotNull] this string argument,
             int length,
-            [CanBeNull, InvokerParameterName] string argumentName = null,
-            [CanBeNull] string message = null)
+            [CanBeNull, InvokerParameterName] string? argumentName = null,
+            [CanBeNull] string? message = null)
         {
             argument.ThrowIfNull(argumentName, message);
 
@@ -293,19 +306,21 @@ namespace FlabIt.Guardians
         /// <param name="message">A custom message that, when specified, will be used instead of the default one.</param>
         /// <exception cref="ArgumentNullException">Raised when <paramref name="argument"/> is null.</exception>
         /// <exception cref="ArgumentLengthLargerThanException">Raised when <paramref name="argument"/> is larger than <paramref name="length"/>.</exception>
+        /// <returns>The <paramref name="argument" /> when it has the same or shorter length than specified with <paramref name="length"/>.</returns>
         [ContractAnnotation("argument:null => halt")]
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static void ThrowIfLargerThan(
-            [NotNull, ValidatedNotNull] this string argument,
+        [NotNull]
+        public static string ThrowIfLargerThan(
+            [System.Diagnostics.CodeAnalysis.NotNull, NotNull, ValidatedNotNull] this string argument,
             int length,
-            [CanBeNull, InvokerParameterName] string argumentName = null,
-            [CanBeNull] string message = null)
+            [CanBeNull, InvokerParameterName] string? argumentName = null,
+            [CanBeNull] string? message = null)
         {
-            argument.ThrowIfNull(argumentName, message);
+            argument = argument.ThrowIfNull(argumentName, message);
 
             if (!IsLargerThan(argument, length))
-                return;
+                return argument;
 
             argumentName ??= nameof(argument);
 
@@ -326,11 +341,12 @@ namespace FlabIt.Guardians
         [DebuggerStepThrough]
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         [NotNull]
+        [Obsolete("Will be replaced by 'ThrowIfLargerThan()' in upcoming versions.")]
         public static string PassThroughNonLargerThan(
-            [NotNull, ValidatedNotNull] this string argument,
+            [System.Diagnostics.CodeAnalysis.NotNull, NotNull, ValidatedNotNull] this string argument,
             int length,
-            [CanBeNull, InvokerParameterName] string argumentName = null,
-            [CanBeNull] string message = null)
+            [CanBeNull, InvokerParameterName] string? argumentName = null,
+            [CanBeNull] string? message = null)
         {
             argument.ThrowIfNull(argumentName, message);