+++ RELEASE_NOTES.md
@@ -1,5 +1,13 @@
 # Release Notes
 
+## v0.8
+
+### Features
+
+* Enabled nullable reference types
+* Marked PassThrough... guardians as obsolete as their ThrowIf... counterpart now is able to return validated input arguments.  
+**Note**: All PassThrough... guardians will be removed in upcoming versions!
+
 ## v0.7
 
 ### Features